+++ src/rx-storage-helper.ts
@@ -168,6 +168,13 @@ export function categorizeBulkWriteRows<RxDocType>(
      * RxStorageInstance().getChangedDocumentsSince().
      */
     changedDocumentIds: RxDocumentData<RxDocType>[keyof RxDocType][];
+
+    /**
+     * TODO directly return a docId->error object
+     * like in the return value of bulkWrite().
+     * This will improve performance because we do not have to iterate
+     * over the error array again.
+     */
     errors: RxStorageBulkWriteError<RxDocType>[];
     eventBulk: EventBulk<RxStorageChangeEvent<RxDocumentData<RxDocType>>>;
 } {